[Sofa.Core] Refactor ObjectFactory
